Add Cart-Order

+++ app/src/main/java/com/ngochuy/ecommerce/data/Product.kt
@@ -21,7 +21,7 @@ data class Product(
         @SerializedName("name")
         val name: String? = null,
         @SerializedName("price")
-        val price: Float? = null,
+        val price: Long? = null,
         @SerializedName("quantity")
         val quantity: Int? = null,
         @SerializedName("quantityOder")
